viator-llm-prototype/requirements.txt
httpx==0.28.1
langchain-core==0.3.84
langchain-openai==0.3.35
langgraph==0.6.11
python-dotenv==1.2.1
pydantic==2.13.3
